refactor(ai): extract pending-event revision helpers (U4)
